fix(package.json/electron/dev-dependencies-to-dependencies): move
Electron can only be a part of dev dependencies. https://github.com/electron-userland/electron-builder/issues/7191. It was moved to dependencies to handle an eslint warning. However, it seems the only fix to the ESLint warning is to disable it.

BREAKING CHANGE: Electron Make now works. It failed on GitHub CI. As a result Electron is back to being a dev dependency thus satisfiying https://github.com/electron-userland/electron-builder/issues/7191 instead of a dev dependency.